==> datasets/enron/3-enron-dataset-experiments.ipynb <==
# Final results for Enron-email dataset
In this notebook, we run the VI algorithm to generate the results for Enron email dataset in the paper.

## Step 1: Import libraries

import pandas as pd
import time
import numpy as np
from datetime import datetime
import torch
from varhawkes.utils import metrics
from tick.hawkes.inference import HawkesADM4
from multiprocessing import Process
import multiprocessing
import bisect

from torch import optim
from varhawkes import models
from varhawkes import posteriors
from varhawkes import priors
from varhawkes import hawkes_model, excitation_kernels
from varhawkes.learners import VariationalInferenceLearner
from experiment.util import LearnerCallback

## Step 2: Read data

df_ts = pd.read_json('enron_dataset_timeseries.json')
df_ts['timestamps'] = df_ts['timestamps'].apply(np.array)
df_ts['timestamps'] = df_ts['timestamps'].apply(np.unique)

# Filter out events before Jan 1, 2000
min_timestamp = time.mktime(datetime(2000,1,1).timetuple())
df_ts['timestamps'] = df_ts['timestamps'].apply(lambda events_m: events_m[events_m > min_timestamp] - min_timestamp)

# Remove the nodes with too few observations (less than 10 emails)
df_ts['num_events_m'] = df_ts['timestamps'].apply(len)
df_ts = df_ts[df_ts.num_events_m > 10] 
df_ts = df_ts.reset_index()

print(df_ts.shape)
df_ts.head()

Scale the time to day-scale

events = list(map(np.array, df_ts.timestamps.values))

scale = 1.0 / (3600 * 24)
end_time = max(map(max, events)) * scale

print('end_time:', end_time)
for m, events_m in enumerate(events):
    events_m = np.unique(np.sort(events_m))
    events_m = events_m * scale
    events[m] = events_m
start_time = min(map(min, events))

Split the data to test & train with proportion 70-30 %

train_thresh = (end_time-start_time) * 0.7 + start_time

def get_events(events, Thr=train_thresh):
    events_train, events_test = list(), list()
    for i, events_i in enumerate(events):
        id_ = bisect.bisect_right(events_i, train_thresh)
        events_train.append(events_i[:id_])
        events_test.append(events_i[id_:]-train_thresh)
    return events_train, events_test

events_train, events_test = get_events(events, Thr=train_thresh)
end_time_train = max([max(nums) for nums in events_train if len(nums)>0])
end_time_test = max([max(nums) for nums in events_test if len(nums)>0])
num_test = sum(map(len,events_test))
print(f'Number of train datapoints: {sum(map(len,events_train))}')
print(f'Number of test datapoints: {num_test}')

---

## Inference with VI-SG

Set the hyper-parameters of the model

M, T = 10, 5
torch.manual_seed(123456789)
np.random.seed(123465789)

Transofrm the data to torch tensor

events_train, events_test = get_events(events, Thr=train_thresh)
end_time_train = max([max(nums) for nums in events_train if len(nums)>0])
end_time_test = max([max(nums) for nums in events_test if len(nums)>0])
events_test = [torch.tensor(events_m, dtype=torch.float64) for events_m in events_test]
events_train = [torch.tensor(events_m, dtype=torch.float64) for events_m in events_train]

#### Train the VI-SG model

hawkes_model_obj = hawkes_model.HawkesModel(excitation_obj, verbose=True)
posterior_obj = posteriors.LogNormalPosterior()
n_params = n_nodes * (n_nodes * M + 1)
C = 1.0 * torch.ones(n_params, dtype=torch.float64)
prior_obj = priors.GaussianLaplacianPrior(n_nodes, n_params, C)
model = models.ModelHawkesVariational(
    model=hawkes_model_obj, posterior=posterior_obj, prior=prior_obj, n_samples=1, n_weights=1, weight_temp=1e0,
)
x = torch.tensor(
    np.hstack((
        np.hstack((  # alpha, the mean of the parameters
            np.random.normal(loc=0.1, scale=0.1, size=n_nodes),
            np.random.normal(loc=0, scale=0.1, size=M * n_nodes ** 2),)),
        np.hstack((  # beta=log(sigma), log of the variance of the parameters    
            np.log(np.clip(np.random.normal(loc=1.0, scale=0.1, size=n_nodes), 1e-1, 2.0)),
            np.log(np.clip(np.random.normal(loc=1.0, scale=0.1, size=M * n_nodes ** 2), 1e-1, 2.0)),))
    )),
    dtype=torch.float64, requires_grad=True
)

learner = VariationalInferenceLearner(model=model, optimizer=optim.Adam([x], lr=0.02), tol=1e-4, max_iter=15e3,
                                      hyperparam_momentum=0.5, hyperparam_interval=100,hyperparam_offset=0)
learner.fit(events_train, end_time_train, x, callback=callback)

### Test the model
Cache the test data

excitation_obj = excitation_kernels.MixtureGaussianFilter(M=M, end_time=T, cut_off=1500.0)
model_likelihood_SG = hawkes_model.HawkesModel(excitation_obj,verbose=True)
model_likelihood_SG.set_data(events_test, end_time=end_time_test)

Testing the model on test data using the mean of approximate posterior

alpha, beta = learner.coeffs.data[:n_params], learner.coeffs.data[n_params:]
z = learner.model.posterior.mean(alpha, beta)
mu_vi = z[:n_nodes].cpu()
adj_vi = z[n_nodes:].cpu().reshape(n_nodes, n_nodes, learner.model.model.excitation.M)
test_likelihood = model_likelihood_SG.log_likelihood(mu_vi,adj_vi)

print(f'M : {M} | T : {T} | log_likelihood (mean) : {test_likelihood/ num_test}')

Testing the model on test data using the mode of approximate posterior

z = learner.model.posterior.mode(alpha, beta)
mu_vi = z[:n_nodes].cpu()
adj_vi = z[n_nodes:].cpu().reshape(n_nodes, n_nodes, learner.model.model.excitation.M)

test_likelihood = model_likelihood_SG.log_likelihood(mu_vi,adj_vi)
print(f'M : {M} | T : {T} | log_likelihood (mode): {test_likelihood/ num_test}')

----

## Inference with VI-EXP

Set the hyper-parameters of the model

M = 1
decay = 20

Cache the testing data

excitation_obj = excitation_kernels.ExponentialKernel(decay=decay, cut_off=1500.0)
model_likelihood = hawkes_model.HawkesModel(excitation_obj,verbose=True)
model_likelihood.set_data(events_test, end_time=end_time_train)

Train the model

excitation_obj = excitation_kernels.ExponentialKernel(decay=decay, cut_off=1500.0)
hawkes_model_obj = hawkes_model.HawkesModel(excitation_obj, verbose=True)
posterior_obj = posteriors.LogNormalPosterior()
n_params = n_nodes * (n_nodes * M + 1)
C = 1.0 * torch.ones(n_params, dtype=torch.float64)
prior_obj = priors.GaussianLaplacianPrior(n_nodes, n_params, C)

model = models.ModelHawkesVariational(
    model=hawkes_model_obj, posterior=posterior_obj, prior=prior_obj, n_samples=1, n_weights=1, weight_temp=1e0,
)
x = torch.tensor(
    np.hstack((
        np.hstack((  # alpha, the mean of the parameters
            np.random.normal(loc=0.1, scale=0.1, size=n_nodes),
            np.random.normal(loc=0, scale=0.1, size=M * n_nodes ** 2),)),
        np.hstack((  # beta=log(sigma), log of the variance of the parameters    
            np.log(np.clip(np.random.normal(loc=1.0, scale=0.1, size=n_nodes), 1e-1, 2.0)),
            np.log(np.clip(np.random.normal(loc=1.0, scale=0.1, size=M * n_nodes ** 2), 1e-1, 2.0)),))
    )),
    dtype=torch.float64, requires_grad=True
)

learner = VariationalInferenceLearner(model=model, optimizer=optim.Adam([x], lr=0.02), tol=1e-4, max_iter=2e4,
                                      hyperparam_momentum=0.5, hyperparam_interval=100,hyperparam_offset=0)
learner.fit(events_train, end_time_train, x)

Testing the model on test data using the mode of approximate posterior

alpha, beta = learner.coeffs.data[:n_params], learner.coeffs.data[n_params:]
z = learner.model.posterior.mode(alpha, beta)
mu_vi = z[:n_nodes].cpu()
adj_vi = z[n_nodes:].cpu().reshape(n_nodes, n_nodes, learner.model.model.excitation.M)

test_likelihood = model_likelihood.log_likelihood(mu_vi,adj_vi)
print(f'Test Likelihood of VI (mode of the posterior): {test_likelihood/num_test}')

Testing the model on test data using the mean of approximate posterior

alpha, beta = learner.coeffs.data[:n_params], learner.coeffs.data[n_params:]
z = learner.model.posterior.mean(alpha, beta)
mu_vi = z[:n_nodes].cpu()
adj_vi = z[n_nodes:].cpu().reshape(n_nodes, n_nodes, learner.model.model.excitation.M)

test_likelihood = model_likelihood.log_likelihood(mu_vi,adj_vi)
print(f'Test Likelihood of VI (mean of the posterior): {test_likelihood/num_test}')
